+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".LaunchScreenActivity">
+        <activity android:name=".Activity.LaunchScreenActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".Activity.MainActivity"
             android:theme="@style/NoActionBarTheme" />
 
         <!-- Dropbox AuthActivity -->
@@ -42,7 +42,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".FileListViewActivity"
+            android:name=".Activity.FileListViewActivity"
             android:theme="@style/TitleTheme" />
     </application>
