+++ app/src/main/java/com/example/kamalksharma/filemanager/Activity/LaunchScreenActivity.java
@@ -1,0 +1,28 @@
+package com.example.kamalksharma.filemanager.Activity;
+
+import android.content.Intent;
+import android.support.v7.app.AppCompatActivity;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+
+import com.example.kamalksharma.filemanager.R;
+
+public class LaunchScreenActivity extends AppCompatActivity  {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_launch_screen);
+        Button mainActivityBtn = (Button) findViewById(R.id.azureBtn);
+        mainActivityBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent startIntent = new Intent(getApplicationContext(), MainActivity.class);
+                startActivity(startIntent);
+            }
+        });
+    }
+
+}
