+++ app/src/main/java/com/example/kamalksharma/filemanager/Utils/UserAccountTask.java
@@ -1,4 +1,4 @@
-package com.example.kamalksharma.filemanager;
+package com.example.kamalksharma.filemanager.Utils;
 
 import android.os.AsyncTask;
 
@@ -11,18 +11,20 @@ import com.dropbox.core.v2.users.FullAccount;
 
 import java.io.File;
 
-public class UserAccountTask extends AsyncTask <Void, Void, ListFolderResult>{
+public class UserAccountTask extends AsyncTask<Void, Void, ListFolderResult> {
     private static final String ACCESS_TOKEN = "Qzmg3GEhnsAAAAAAAAAEs05bMlnYeXIclE1nFUyF1-nfnFVhCXPpvaTCdF0EU94n";
     private DbxClientV2 dbxClient;
-    private TaskDelegate  delegate;
+    private TaskDelegate delegate;
     private Exception error;
     private String mpath;
+
     public interface TaskDelegate {
         void onAccountReceived(ListFolderResult account);
+
         void onError(Exception error);
     }
 
-    public UserAccountTask(String mpath, TaskDelegate delegate){
+    public UserAccountTask(String mpath, TaskDelegate delegate) {
         this.mpath = mpath;
         DbxRequestConfig config = DbxRequestConfig.newBuilder("dropbox/KakesApp").build();
         DbxClientV2 client = new DbxClientV2(config, ACCESS_TOKEN);
@@ -47,11 +49,10 @@ public class UserAccountTask extends AsyncTask <Void, Void, ListFolderResult>{
     protected void onPostExecute(ListFolderResult account) {
         super.onPostExecute(account);
 
-        if (account != null && error == null){
+        if (account != null && error == null) {
             //User Account received successfully
             delegate.onAccountReceived(account);
-        }
-        else {
+        } else {
             // Something went wrong
             delegate.onError(error);
         }
