+++ app/src/main/res/layout/list_fileitem_layout.xml
@@ -1,20 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="horizontal" android:layout_width="match_parent"
+    android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:orientation="horizontal"
     android:padding="8dp">
 
     <ImageView
         android:id="@+id/fileIcon"
         android:layout_width="50dp"
         android:layout_height="50dp"
-        android:layout_alignParentLeft="true"
         android:layout_alignParentStart="true"
+        android:layout_alignParentLeft="true"
         android:layout_centerVertical="true"
-        android:layout_marginLeft="16dp"
         android:layout_marginStart="0dp"
-        android:src="@drawable/folder"
-        />
+        android:layout_marginLeft="16dp"
+        android:src="@drawable/folder" />
 
     <TextView
         android:id="@+id/fileName"
@@ -24,16 +24,14 @@
         android:layout_weight="1"
         android:padding="16dp"
         android:text="@string/app_name"
-        android:textSize="12dp"
-        />
+        android:textSize="12dp" />
+
     <ImageView
         android:id="@+id/folderArrowIcon"
         android:layout_width="20dp"
         android:layout_height="40dp"
-        android:text="sadasdasda"
         android:layout_gravity="center_vertical"
-       />
-
+        android:text="sadasdasda" />
 
 
 </LinearLayout>
